react study - management
