--- myCode/client/README.md
@@ -1,21 +1,0 @@
-# React Customer Management System
-
-- React Customer Management System 학습하기
-
-- 기본적으로 리액트는 html 문서를 어떻게 효과적으로 보여줄 것인지에 대해 많은 고민을 한 결과물로 만들어진 라이브러리다.
-
-- 그리고 html 문서를 한 화면에서 잘게 쪼개는데 그것이 컴포넌트다 (Component)
-
-## props와 state의 처리 시, 차이점
-
-- props는 일반적으로 변경할 수 없는 데이터를 처리할 때 사용하고
-
-- state는 변경이 가능한 데이터를 처리할 때 사용한다.
-
-## 2회 학습 시작
-
-- 강의 자체가 기능 구현에 초점이 맞춰진 프로젝트라 빠르게 다시 한 번 볼 수 있을 것 같다
-
-- 한 번 강의를 돌리고 나니, 리액트 이해가 좀 더 쉬워졌다. 21일 전까지 프로젝트 마무리하자.
-
-- Javascript에서 Static 메소드가 활용되는 이유
