--- myCode/client/src/App.css
@@ -1,3 +1,0 @@
-.gray-background {
-  background-color: darkgray;
-}
