+++ myCode/management/client/src/components/Customer.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import React, {Component} from 'react'
 import TableRow from '@material-ui/core/TableRow'
 import TableCell from '@material-ui/core/TableCell'
 
@@ -18,28 +18,5 @@ class Customer extends React.Component{
         )
     }
 }
-
-// class CustomerProfile extends React.Component{
-//     render(){
-//         return(
-//             <div>
-//                 <img src={this.props.image} alt="profile"/>
-//         <h2>{this.props.name} ({this.props.id})</h2>
-//             </div>
-//         )
-//     }
-// }
-
-// class CustomerInfo extends React.Component{
-//     render(){
-//         return (
-//             <div>
-//             <p>{this.props.age}</p>
-//             <p>{this.props.gender}</p>
-//             <p>{this.props.job}</p>
-//             </div>
-//         )
-//     }
-// }
 
 export default Customer
